requirements.txt
requests
pygame
python-dotenv